% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prox}
\alias{prox}
\title{proximal function}
\usage{
prox(x, alpha, func = c("l0", "l1", "l0proj"))
}
\description{
objective = f(w)+g(w);
g(w) = alpha*norm(w)_0(func = "l0"), alpha*norm(w)_1 (func="l1") or I_(norm(w)_0<=alpha)(w) (func="l0proj")
}
