\name{simPU-package}
\alias{simPU-package}
\alias{simPU}
\docType{package}
\title{
\packageTitle{simPU}
}
\description{
\packageDescription{simPU}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simPU}
\packageIndices{simPU}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{simPU}

Maintainer: \packageMaintainer{simPU}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
